# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Patchman-engine API","description":"Description here","contact":{},"license":{},"version":"1.0"},"servers":[{"url":"/"}],"paths":{"/api/patch/v1/advisories":{"get":{"summary":"Show me all applicable errata for all my systems","description":"Show me all applicable errata for all my systems","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/controllers.AdvisoriesResponse"}}}}},"security":[{"RhIdentity":[]}]}},"/api/patch/v1/advisories/{advisory_id}":{"get":{"summary":"Show me details an advisory by given advisory name","description":"Show me details an advisory by given advisory name","parameters":[{"name":"advisory_id","in":"path","description":"Advisory ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/controllers.AdvisoryDetailResponse"}}}}},"security":[{"RhIdentity":[]}]}},"/api/patch/v1/advisories/{advisory_id}/systems":{"get":{"summary":"Show me systems on which the given advisory is applicable","description":"Show me systems on which the given advisory is applicable","parameters":[{"name":"advisory_id","in":"path","description":"Advisory ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/controllers.AdvisorySystemsResponse"}}}}},"security":[{"RhIdentity":[]}]}},"/api/patch/v1/systems":{"get":{"summary":"Show me all my systems","description":"Show me all my systems","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/controllers.SystemsResponse"}}}}},"security":[{"RhIdentity":[]}]}},"/api/patch/v1/systems/{inventory_id}":{"get":{"summary":"Show me details about a system by given inventory id","description":"Show me details about a system by given inventory id","parameters":[{"name":"inventory_id","in":"path","description":"Inventory ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/controllers.SystemDetailResponse"}}}}},"security":[{"RhIdentity":[]}]}},"/api/patch/v1/systems/{inventory_id}/advisories":{"get":{"summary":"Show me advisories for a system by given inventory id","description":"Show me advisories for a system by given inventory id","parameters":[{"name":"inventory_id","in":"path","description":"Inventory ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/controllers.SystemAdvisoriesResponse"}}}}},"security":[{"RhIdentity":[]}]}}},"components":{"schemas":{"controllers.AdvisoriesResponse":{"type":"object","properties":{"data":{"type":"array","description":"advisories items","items":{"$ref":"#/components/schemas/controllers.AdvisoryItem"}},"links":{"$ref":"#/components/schemas/controllers.Links"},"meta":{"$ref":"#/components/schemas/controllers.AdvisoryMeta"}}},"controllers.AdvisoryDetailAttributes":{"type":"object","properties":{"cves":{"type":"array","items":{"type":"string"}},"description":{"type":"string"},"fixes":{"type":"string"},"modified_date":{"type":"string"},"public_date":{"type":"string"},"references":{"type":"array","items":{"type":"string"}},"severity":{"type":"string"},"solution":{"type":"string"},"synopsis":{"type":"string"},"topic":{"type":"string"}}},"controllers.AdvisoryDetailItem":{"type":"object","properties":{"attributes":{"$ref":"#/components/schemas/controllers.AdvisoryDetailAttributes"},"id":{"type":"string"},"type":{"type":"string"}}},"controllers.AdvisoryDetailResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/controllers.AdvisoryDetailItem"}}},"controllers.AdvisoryItem":{"type":"object","properties":{"attributes":{"$ref":"#/components/schemas/controllers.AdvisoryItemAttributes"},"id":{"type":"string"},"type":{"type":"string"}}},"controllers.AdvisoryItemAttributes":{"type":"object","properties":{"advisory_type":{"type":"integer"},"applicable_systems":{"type":"integer"},"description":{"type":"string"},"public_date":{"type":"string"},"severity":{"type":"string"},"synopsis":{"type":"string"}}},"controllers.AdvisoryMeta":{"type":"object","properties":{"data_format":{"type":"string"},"filter":{"type":"string"},"limit":{"type":"integer"},"offset":{"type":"integer"},"page":{"type":"integer"},"page_size":{"type":"integer"},"pages":{"type":"integer"},"public_from":{"type":"integer"},"public_to":{"type":"integer"},"severity":{"type":"string"},"show_all":{"type":"boolean"},"sort":{"type":"boolean"},"total_items":{"type":"integer"}}},"controllers.AdvisorySystemsMeta":{"type":"object","properties":{"advisory":{"type":"string"},"data_format":{"type":"string"},"enabled":{"type":"boolean"},"filter":{"type":"string"},"limit":{"type":"integer"},"offset":{"type":"integer"},"page":{"type":"integer"},"page_size":{"type":"integer"},"pages":{"type":"integer"},"total_items":{"type":"integer"}}},"controllers.AdvisorySystemsResponse":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/controllers.SystemItem"}},"links":{"$ref":"#/components/schemas/controllers.Links"},"meta":{"$ref":"#/components/schemas/controllers.AdvisorySystemsMeta"}}},"controllers.Links":{"type":"object","properties":{"first":{"type":"string"},"last":{"type":"string"},"next":{"type":"string"},"previous":{"type":"string"}}},"controllers.SystemAdvisoriesResponse":{"type":"object","properties":{"data":{"type":"array","description":"advisories items","items":{"$ref":"#/components/schemas/controllers.AdvisoryItem"}},"links":{"$ref":"#/components/schemas/controllers.Links"},"meta":{"$ref":"#/components/schemas/controllers.AdvisoryMeta"}}},"controllers.SystemDetailResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/controllers.SystemItem"}}},"controllers.SystemItem":{"type":"object","properties":{"attributes":{"$ref":"#/components/schemas/controllers.SystemItemAttributes"},"id":{"type":"string"},"type":{"type":"string"}}},"controllers.SystemItemAttributes":{"type":"object","properties":{"enabled":{"type":"boolean"},"last_evaluation":{"type":"string"},"last_upload":{"type":"string"},"rhba_count":{"type":"integer"},"rhea_count":{"type":"integer"},"rhsa_count":{"type":"integer"}}},"controllers.SystemsMeta":{"type":"object","properties":{"data_format":{"type":"string"},"enabled":{"type":"boolean"},"filter":{"type":"string"},"limit":{"type":"integer"},"offset":{"type":"integer"},"page":{"type":"integer"},"page_size":{"type":"integer"},"pages":{"type":"integer"},"total_items":{"type":"integer"}}},"controllers.SystemsResponse":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/controllers.SystemItem"}},"links":{"$ref":"#/components/schemas/controllers.Links"},"meta":{"$ref":"#/components/schemas/controllers.SystemsMeta"}}}},"securitySchemes":{"RhIdentity":{"type":"apiKey","name":"x-rh-identity","in":"header"}}}}